refactor: enum 수정
