--- src/main/java/com/example/dropshop/domain/refund/entity/RefundStatus.java
@@ -1,11 +1,0 @@
-package com.example.dropshop.domain.refund.entity;
-
-/** 환불 상태. */
-public enum RefundStatus {
-  PENDING,
-  APPROVED,
-  PROCESSING,
-  COMPLETED,
-  FAILED,
-  REJECTED;
-}
